+++ integrated_channels/sap_success_factors/admin/__init__.py
@@ -5,10 +5,13 @@ from __future__ import absolute_import, unicode_literals
 
 from django.contrib import admin
 from config_models.admin import ConfigurationModelAdmin
+from requests import RequestException
+from simplejson import JSONDecodeError
 
 from integrated_channels.sap_success_factors.models import (
     SAPSuccessFactorsEnterpriseCustomerConfiguration, SAPSuccessFactorsGlobalConfiguration
 )
+from integrated_channels.sap_success_factors.client import SAPSuccessFactorsAPIClient
 
 
 @admin.register(SAPSuccessFactorsGlobalConfiguration)
@@ -32,15 +35,21 @@ class SAPSuccessFactorsEnterpriseCustomerConfigurationAdmin(admin.ModelAdmin):
     """
     Django admin model for SAPSuccessFactorsEnterpriseCustomerConfiguration.
     """
+
     list_display = (
         "enterprise_customer_name",
         "active",
         "sapsf_base_url",
         "key",
+        "secret",
+        "sapsf_company_id",
+        "sapsf_user_id",
+        "has_access_token",
     )
 
     readonly_fields = (
         "enterprise_customer_name",
+        "has_access_token",
     )
 
     list_filter = ("active",)
@@ -51,10 +60,34 @@ class SAPSuccessFactorsEnterpriseCustomerConfigurationAdmin(admin.ModelAdmin):
 
     def enterprise_customer_name(self, obj):
         """
-        Returns: the name for the attached EnterpriseCustomer, or the uuid if the EnterpriseCustomer does not exist.
+        Returns: the name for the attached EnterpriseCustomer.
 
         Args:
             obj: The instance of SAPSuccessFactorsEnterpriseCustomerConfiguration
                 being rendered with this admin form.
         """
         return obj.enterprise_customer.name
+
+    def has_access_token(self, obj):
+        """
+        Confirms the presence and validity of the access token for the SAP SuccessFactors client instance
+
+        Returns: a bool value indicating the presence of the access token
+
+        Args:
+            obj: The instance of SAPSuccessFactorsEnterpriseCustomerConfiguration
+                being rendered with this admin form.
+        """
+        try:
+            access_token, expires_at = SAPSuccessFactorsAPIClient.get_oauth_access_token(
+                obj.sapsf_base_url,
+                obj.key,
+                obj.secret,
+                obj.sapsf_company_id,
+                obj.sapsf_user_id
+            )
+        except (RequestException, JSONDecodeError):
+            return False
+        return bool(access_token and expires_at)
+
+    has_access_token.boolean = True