+++ integrated_channels/sap_success_factors/models.py
@@ -50,8 +50,10 @@ class SAPSuccessFactorsEnterpriseCustomerConfiguration(EnterpriseCustomerPluginC
     The Enterprise specific configuration we need for integrating with SuccessFactors.
     """
 
-    sapsf_base_url = models.CharField(max_length=255)
     key = models.CharField(max_length=255, blank=True, verbose_name="Client ID")
+    sapsf_base_url = models.CharField(max_length=255, verbose_name="SAP Base URL")
+    sapsf_company_id = models.CharField(max_length=255, blank=True, verbose_name="SAP Company ID")
+    sapsf_user_id = models.CharField(max_length=255, blank=True, verbose_name="SAP User ID")
     secret = models.CharField(max_length=255, blank=True, verbose_name="Client Secret")
 
     history = HistoricalRecords()